setup grpc
bot is semi setup, no time tonight to do testing.

a few functions still need to be setup. 
GRPC proto file setup and generated the file for use within and on other clients. 

had to move main file to within a folder to make folder structure easy for external services to utilise the generated file.